+++ src/RdfaLiteMicrodata/Tests/Infrastructure/HtmlDocumentFactoryTest.php
@@ -86,4 +86,17 @@ class HtmlDocumentFactoryTest extends \PHPUnit_Framework_TestCase
             $this->assertEquals('Tag invalid invalid', trim($parsingError->message));
         }
     }
+
+    /**
+     * Test a custom HTML parsing error handler
+     */
+    public function testHtmlDocumentCustomErrorHandler()
+    {
+        $customErrorHandler = function(\LibXMLError $error) {
+            return ($error->level == 2) && ($error->code == 801) && (trim($error->message) == 'Tag invalid invalid');
+        };
+        $htmlDocumentFactory = new HtmlDocumentFactory($customErrorHandler);
+        $htmlSource = '<html><head><title>Test</title></head><body><invalid>Test</invalid></body></html>';
+        $htmlDocumentFactory->createDocumentFromSource($htmlSource);
+    }
 }