+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All Notable changes to *jkphl/rdfa-lite-microdata* will be documented in this fi
 ## [0.3.1] - 2017-03-24
 ### Added
 * Dedicated exception for HTML parsing errors
+* Added custom HTML parsing error handling ([#4](https://github.com/jkphl/rdfa-lite-microdata/issues/4))
 
 ## [0.3.0] - 2017-03-17
 ### Added